requirements.txt
discord.py==2.1.0
requests==2.28.1
python-dotenv==0.21.0
youtube_dl